AUTH [#258] Centralize cookie forwarding logic into a shared helper
